refactor: ConsoleUtil 분리
